Try explicit cast to long on line number for older expat libs that may not have the routine defined that way.

+++ ccnd/lib/ccn_xmltoccnb.c
@@ -401,7 +401,7 @@ process_fd(int fd, FILE *outfile, int flags)
         res |= 1;
     }
     if (XML_Parse(p, buf, 0, 1) != XML_STATUS_OK) {
-      fprintf(stderr, "xml parse error line %ld\n", XML_GetCurrentLineNumber(p));
+      fprintf(stderr, "xml parse error line %ld\n", (long)XML_GetCurrentLineNumber(p));
         res |= 1;
     }
     XML_ParserFree(p);
